Use flask blueprints instead of controllers
