+++ ckanext/tiledmap/routes/__init__.py
@@ -1,0 +1,9 @@
+#!/usr/bin/env python
+# encoding: utf-8
+#
+# This file is part of ckanext-versioned-tiledmap
+# Created by the Natural History Museum in London, UK
+
+from . import map
+
+blueprints = [map.blueprint]
